changes on how and where to add users
i changed it so the when we fetch the users we get the ones in the db not the file so user.json is now useless.

i tried to make it so when we add a user it adds to the db,but for some fkin reason it keeps redirecting me to the welcome.blade.php which i have no fkin idea why like no line of code indicates it should do that plz help.