Look, I tried. I really did.
I tried adding the events that shows the user when the driver receives their order and is delivering it. I'm trying to use Laravel Pusher, but so far it hasn't worked for me. Here's a link to the website, sign up to it, so i can add you as a collaborator to the NovaCart app: https://pusher.com