Did a BUNCH of shit
Added Models, Migrations, Factories and Controllers for Orders and Drivers. Changed the seeder so random users have isDriver set to true, any user that isDriver creates a Driver instance with the same username, location and userID. Everytime a user confirms a purchase, a new instance of the order model is created that has a 'content' variable(products bought) userID for who bought them and isAccepted boolean. Tried adding a timer to cycle through the states of the order, but it doesn't COMPLETELY work (check AppServiceProvider.php and UpdateNotifications.php.   Try and see if you can make it work, maybe the problems lie with Postman?) (ignore OrderShipmentStatusUpdated.php, broadcasting.php, reverb.php, echo.js, and channels.php. Those were for the broadcasting crap that didn't work.)