Did some confusing shite
Now, when calling the /notif method, the notification will cycle through [null, pending, accepted, delivering, delivered] states, it also changes some variables like isAccepted for a user, isAccepted for an order, isDelivering for a driver, the order also gets deleted after being delivered.
Added JSON responses to a few of the methods, so the frontend can easily figure out if the method worked or not. Failing to sign up will also return the variables that are already in use by other users (for example if someone signed up with an email and number already used by another user, it will return those fields, so the frontend can easily highlight what to change).
Ignore the migration file being removed and added again, I kinda fucked up for a small bit. lol