Good few changes
Now when the user logs in unsuccessfully, it will show why it failed, either user not found(enter number not in users) or email/number mismatch (number is correct but email isn't) or password is incorrect (number and email are correct, so the password must be incorrect), also added a case or two for when we enter only the number and password.
Also, when the user sends a product ID for the favourite thing but it's already in favs, it will remove it, a small little shortcut instead of adding a whole new function.
Made the driver_id in the order migration nullable, as the order will be created before the driver confirms the delivery, also we're using the driver_id in the orders so when the user logs in, he will see WHO is the driver delivering his order.
Fixed the issue where after modifying an array, it suddenly becomes numbered(check UserController and CartController)
Only a few changes left we need to make that i can think of on the top of my head.