Started working on the website, plus some minor changes
Now there's a website to show the products and the stores, and a website that makes you choose which when you want to see/edit
We have an edit and delete button for each product.
Now we just need to make routes to actually edit the database.
I'll text you what to do