Pictures, logos, fonts.... oh my
Configured the image/logo attribute to the users, products and stores. Now they point to the folder in the project. Each user now has a default logo(in Storage/app/public/users), and each time changelogo is called, it adds a new picture file to the Users folder.
Also used a new font, "Forte Regular" to write in some web pages.
There's a new ImageController that shows the picture you're looking for(depending on the path and image name), I don't think we need it, but it's good to have it just to check that the pictures are working properly.