Cleaned the code, we're pretty much done now
Added a redirect page for the update and delete functions now.
Fixed up all the problems in the driver pages, and polished the code, so it works properly.'
Editing and adding stores and drivers with credentials that are NOT unique(like the name, email or number) will return an error.
Info changed in a driver, will change in the corresponding user.
Fixed an error where changing the the image of a driver that had a default image will delete the default image completely.
All we need to do now is QA testing.
Good job, brother!